% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Computes entropy of matrix}
\usage{
entropy(mu)
}
\arguments{
\item{mu}{a symmetric matrix with zeroes in the diagonal}
}
\description{
this function computes the entropy of (symmetric) matrix \code{mu}. The input should be
a \code{n}*\code{n} symmetric matrix, with zeros in the diagonal.
}
\examples{
mu <- matrix(c(0,0.2,0.2,0.2,0,0.2,0.2,0.2,0), 
           nrow = 3, 
           ncol = 3
           )
entropy(mu)


}
\author{
Angelo Mele
}

