% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.network.R
\name{initialize.network}
\alias{initialize.network}
\title{generate initial network as random erdos-renyi graph}
\usage{
initialize.network(theta, n, directed = FALSE)
}
\arguments{
\item{theta}{vector of parameters for the ergm}

\item{n}{size of the network}

\item{directed}{logical: FALSE if undirected network (default is FALSE)}
}
\description{
this function generates the initial network, to start
the simulations of the ergm, as an erdos-renyi graph
}
\examples{
n <- 10
theta <- c(-1,3)
g <- initilize.network(theta, n, directed = FALSE)
g


}
\author{
Angelo Mele
}

