% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logconstant_mf2.R
\name{logconstant_mf2}
\alias{logconstant_mf2}
\title{Compute log constant of ERGM using mean-field approximation}
\usage{
logconstant_mf2(pp, mu, theta, x, dt, tol, maxiterations = 3)
}
\arguments{
\item{pp}{2-dimensional array. the first entry 
indicates how many variables are in the direct utility. 
The second entry counts how many parameters.}

\item{mu}{matrix of initial values for mu}

\item{theta}{parameters of the ergm}

\item{x}{matrix with individual nodes attributes}

\item{dt}{matrix with indicator of which column of x and which change statistics to use}

\item{tol}{tolerance for the iterative process}

\item{maxiterations}{how many iterations of the variational approximation}
}
\description{
This function computes the log-constant of the ERGM
using a mean-field approximation of the model
}
\examples{
n <- 10   # small network
x <- rnorm(n,0,1)  #attributes
mu <- matrix(runif(n^2,0,1), nrow = n, ncol = n) #initialize matrix mu
diag(mu) <- 0  #make sure mu has 0 diagonal
# Now let's get the variational mean-field aprpoximation
p <- c(1,2)  # 2 variables only
q <- 1   # x has only 1 column
dt <- matrix(0, ncol = p[2], nrow = 2) 
dt[,1] <- c(1,1)  # which column of x to consider 
dt[,2] <- c(1,1)  # which change stats to use
theta = c(-3.0,1.5)  # parameters
tol = 0.000001   # tolerance
psi <- logconstant_mf2(p, mu, theta, x, dt, tol)  #compute log-constant
psi$psi  
}

