% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfcd.R
\name{mfcd}
\alias{mfcd}
\title{solves Chatterjee-Diaconis 2013 variational problem for model with homogeneous players and
only edges and 2-stars}
\usage{
mfcd(a, b)
}
\arguments{
\item{a}{parameter for edges term}

\item{b}{parameter for 2-stars term}
}
\description{
this function computes solutions to Chatterjee-Diaconis 2013 variational problem.
Returns a matrix with solution (row 1) and first derivative at the solution (row 2)
}
\examples{
mfcd(a = -2, b = 0.2)


}
\author{
Angelo Mele
}

