\name{mfergm-package}
\alias{mfergm-package}
\alias{mfergm}
\docType{package}
\title{
Mean-Field variational approximation for ERGM models

}
\description{
mfergm is a package for estimation of ergm using a mean-field variational 
approximation algorithm. 
}
\details{
\tabular{ll}{
Package: \tab mfergm\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-12\cr
License: \tab GPL\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Angelo Mele <angelo.mele@jhu.edu>

Maintainer: Angelo Mele <angelo.mele@jhu.edu>
}
\references{
Mele and Zhu (2015) 
}
