% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_mfcd.R
\name{phi_mfcd}
\alias{phi_mfcd}
\title{mean-field function for the Chatterjee-Disconis variational problem}
\usage{
phi_mfcd(p, a, b)
}
\arguments{
\item{p}{prob of link (mean field approximation)}

\item{a}{parameter for edges term}

\item{b}{parameter for 2-stars term}
}
\description{
this function is the same as Chatterjee-Diaconis first order conditions
}
\examples{
phi_mfcd(p = 0.2, a = -2, b = 0.2)


}
\author{
Angelo Mele
}

