% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_prime_mfcd.R
\name{phi_prime_mfcd}
\alias{phi_prime_mfcd}
\title{second derivative of variational problem in Chatterjee-Diaconis 2013}
\usage{
phi_prime_mfcd(p, a, b)
}
\arguments{
\item{p}{prob of link (mean field approximation)}

\item{a}{parameter for edges term}

\item{b}{parameter for 2-stars term}
}
\description{
this function is the second derivative of the variational problem
}
\examples{
phi_prime_mfcd(p = 0.2, a = -2, b = 0.2)


}
\author{
Angelo Mele
}

