% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.model2.R
\name{simulate.model2}
\alias{simulate.model2}
\title{Generate networks and estimates parameters}
\usage{
\method{simulate}{model2}(theta, n = 10, nsims = 2, ninit = 5,
  ergm = FALSE, cd = TRUE, mfergm = FALSE, mple = FALSE,
  sim.seed = 1977)
}
\arguments{
\item{theta}{true parameters vector}

\item{n}{size of the network}

\item{nsims}{number of simulated networks to estimate}

\item{ninit}{number of trials for mean field approximation (if mfergm = TRUE)}

\item{ergm}{logical, if FALSE will not perform ergm estimation}

\item{cd}{logical, if FALSE will not perform Chatterjee-Diaconis estimation}

\item{mfergm}{logical, if FALSE will not perform mean-field estimation}

\item{mple}{logical, if FALSE will not perform MPLE estimation}

\item{sim.seed}{seed for the random number generator}
}
\description{
This function simulates \code{nsims} networks, and 
compares estimates using 
the ergm routines, the Chatterje-Diaconis 2013 mean-field
and the more general mean-field approximation. The model
is a model with edges and 2-stars.
}

